refactor: move loader
