+++ pkg/symbol/loader.go
@@ -1,4 +1,4 @@
-package loader
+package symbol
 
 import (
 	"context"
@@ -10,27 +10,26 @@ import (
 	"github.com/siyul-park/uniflow/pkg/database"
 	"github.com/siyul-park/uniflow/pkg/spec"
 	"github.com/siyul-park/uniflow/pkg/store"
-	"github.com/siyul-park/uniflow/pkg/symbol"
 )
 
-// Config contains configuration settings for the Loader.
-type Config struct {
-	Namespace string        // Namespace associated with the Loader
-	Table     *symbol.Table // Symbol table for storing loaded symbols
-	Store     *store.Store  // Store to retrieve spec.Spec from
+// LoaderConfig contains configuration settings for the Loader.
+type LoaderConfig struct {
+	Namespace string       // Namespace associated with the Loader
+	Table     *Table       // Symbol table for storing loaded symbols
+	Store     *store.Store // Store to retrieve spec.Spec from
 }
 
-// Loader synchronizes with the store.Store to load spec.Spec into the symbol.Table.
+// Loader synchronizes with the store.Store to load spec.Spec into the Table.
 type Loader struct {
 	namespace string        // Namespace for loading
-	table     *symbol.Table // Symbol table instance
+	table     *Table        // Symbol table instance
 	store     *store.Store  // Store instance
 	stream    *store.Stream // Stream for watching changes
 	mu        sync.RWMutex  // Mutex for synchronization
 }
 
-// New creates a new Loader instance with the given configuration.
-func New(config Config) *Loader {
+// NewLoader creates a new Loader instance with the given configuration.
+func NewLoader(config LoaderConfig) *Loader {
 	return &Loader{
 		namespace: config.Namespace,
 		table:     config.Table,
@@ -39,7 +38,7 @@ func New(config Config) *Loader {
 }
 
 // LoadOne loads a single spec.Spec identified by ID and its linked specs into the symbol table.
-func (l *Loader) LoadOne(ctx context.Context, id uuid.UUID) (*symbol.Symbol, error) {
+func (l *Loader) LoadOne(ctx context.Context, id uuid.UUID) (*Symbol, error) {
 	l.mu.Lock()
 	defer l.mu.Unlock()
 
@@ -130,7 +129,7 @@ func (l *Loader) LoadOne(ctx context.Context, id uuid.UUID) (*symbol.Symbol, err
 }
 
 // LoadAll loads all spec.Spec from the store and adds them to the symbol table.
-func (l *Loader) LoadAll(ctx context.Context) ([]*symbol.Symbol, error) {
+func (l *Loader) LoadAll(ctx context.Context) ([]*Symbol, error) {
 	l.mu.Lock()
 	defer l.mu.Unlock()
 
@@ -154,7 +153,7 @@ func (l *Loader) LoadAll(ctx context.Context) ([]*symbol.Symbol, error) {
 		return nil, err
 	}
 
-	var symbols []*symbol.Symbol
+	var symbols []*Symbol
 	for _, spec := range specs {
 		if sym, err := l.table.Insert(spec); err != nil {
 			return nil, err
